move TabData type to TabRender (#44)
