+++ client/src/route/auth/AccountTabs.tsx
@@ -1,5 +1,6 @@
 import * as React from 'react';
-import TabComponent, { TabData } from 'components/tab/TabComponent';
+import TabComponent from 'components/tab/TabComponent';
+import { TabData } from 'components/tab/subcomponents/TabRender';
 
 const tabs: TabData[] = [
   {
